+++ lib/AppInfo/Application.php
@@ -13,7 +13,6 @@ use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IRegistrationContext;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
-use OCP\Notification\IManager as INotificationManager;
 
 use OCA\SuiteCRM\Dashboard\SuiteCRMWidget;
 use OCA\SuiteCRM\Search\SuiteCRMSearchProvider;
@@ -28,22 +27,14 @@ class Application extends App implements IBootstrap {
 
 	public const APP_ID = 'integration_suitecrm';
 
-	/**
-	 * Constructor
-	 *
-	 * @param array $urlParams
-	 */
 	public function __construct(array $urlParams = []) {
 		parent::__construct(self::APP_ID, $urlParams);
-
-		$container = $this->getContainer();
-		$manager = $container->get(INotificationManager::class);
-		$manager->registerNotifierService(Notifier::class);
 	}
 
 	public function register(IRegistrationContext $context): void {
 		$context->registerDashboardWidget(SuiteCRMWidget::class);
 		$context->registerSearchProvider(SuiteCRMSearchProvider::class);
+		$context->registerNotifierService(Notifier::class);
 	}
 
 	public function boot(IBootContext $context): void {