+++ src/bootstrap.js
@@ -1,7 +1,8 @@
-import Vue from 'vue'
 import { translate, translatePlural } from '@nextcloud/l10n'
 
-Vue.prototype.t = translate
-Vue.prototype.n = translatePlural
-Vue.prototype.OC = window.OC
-Vue.prototype.OCA = window.OCA
+export function applyGlobals(app) {
+	app.config.globalProperties.t = translate
+	app.config.globalProperties.n = translatePlural
+	app.config.globalProperties.OC = window.OC
+	app.config.globalProperties.OCA = window.OCA
+}