+++ src/components/AdminSettings.vue
@@ -63,7 +63,7 @@ export default {
 	components: {
 	},
 
-	props: [],
+	props: {},
 
 	data() {
 		return {