+++ src/components/PersonalSettings.vue
@@ -99,7 +99,7 @@ export default {
 	components: {
 	},
 
-	props: [],
+	props: {},
 
 	data() {
 		return {