+++ src/personalSettings.js
@@ -1,25 +1,7 @@
-/* jshint esversion: 6 */
+import { createApp } from 'vue'
+import { applyGlobals } from './bootstrap.js'
+import PersonalSettings from './components/PersonalSettings.vue'
 
-/**
- * Nextcloud - suitecrm
- *
- *
- * This file is licensed under the Affero General Public License version 3 or
- * later. See the COPYING file.
- *
- * @author Julien Veyssier <eneiluj@posteo.net>
- * @copyright Julien Veyssier 2020
- */
-
-import Vue from 'vue'
-import './bootstrap'
-import PersonalSettings from './components/PersonalSettings'
-
-// eslint-disable-next-line
-'use strict'
-
-// eslint-disable-next-line
-new Vue({
-	el: '#suitecrm_prefs',
-	render: h => h(PersonalSettings),
-})
+const app = createApp(PersonalSettings)
+applyGlobals(app)
+app.mount('#suitecrm_prefs')